added: hex/hex.go API entrypoint; changed: reformatting the functions from record -> hex.
