+++ internal/evaluator/hex/record.go
@@ -3,8 +3,6 @@ package main
 import (
 	"encoding/binary"
 	"encoding/hex"
-	"fmt"
-	"io"
 )
 
 const (
@@ -53,54 +51,6 @@ func (r Record) RecordType() RecordType {
 	return RecordType(rType)
 }
 
-func parseRecord(input io.ByteReader) (Record, error) {
-	var record Record
-
-	curr, err := input.ReadByte()
-	if err != nil {
-		return Record{}, NoMoreRecordsErr
-	}
-
-	if curr != startCode {
-		return Record{}, MissingStartCodeErr
-	}
-
-	for curr != '\r' && curr != '\n' {
-		record = append(record, curr)
-		curr, err = input.ReadByte()
-		if err != nil {
-			return Record{}, WrongRecordFormatErr
-		}
-	}
-
-	if curr == '\r' {
-		curr, err = input.ReadByte()
-		if err != nil || curr != '\n' {
-			return nil, WrongRecordFormatErr
-		}
-	}
-	return record, nil
-}
-
-func validateRecord(rec Record) bool {
-	recordLen := len(rec)
-	if recordLen < minLength {
-		return false
-	}
-
-	dataLenBytes := make([]byte, 2)
-	_, err := hex.Decode(dataLenBytes, rec[1:3])
-	if err != nil {
-		return false
-	}
-
-	dataLen := binary.LittleEndian.Uint16(dataLenBytes)
-	if recordLen != int(minLength+dataLen) {
-		return false
-	}
-	return true
-}
-
 func hexToInt(data []byte, littleEndian bool) (uint64, error) {
 	dataLen := len(data)
 	if dataLen%2 == 1 {
@@ -120,51 +70,4 @@ func hexToInt(data []byte, littleEndian bool) (uint64, error) {
 	}
 
 	return endianness.Uint64(dataLenBytes), nil
-}
-
-type HexFile struct {
-	base    int
-	records []Record
-}
-
-func ReadAll(in io.ByteReader) (*HexFile, error) {
-	// add check for only one eof record
-	var records []Record
-	rec, err := parseRecord(in)
-	for ; err == nil; rec, err = parseRecord(in) {
-		if validateRecord(rec) {
-			records = append(records, rec)
-		}
-	}
-
-	if err == NoMoreRecordsErr {
-		return &HexFile{base: 0, records: records}, nil
-	}
-	return nil, err
-}
-
-func Checksum(record []byte) ([2]byte, error) {
-	var checksum uint64
-	var hexStringChecksum [2]byte
-	start := 0
-	if record[0] == startCode {
-		start = 1
-	}
-
-	decoded := make([]byte, hex.DecodedLen(len(record[start:])))
-	if _, err := hex.Decode(decoded, record[start:]); err != nil {
-		return hexStringChecksum, fmt.Errorf("%s: %w", WrongRecordFormatErr, err)
-	}
-
-	hexChecksum := uint64(0)
-	for _, hexVal := range decoded {
-		hexChecksum += uint64(hexVal)
-	}
-
-	checksumCont := make([]byte, 8)
-	checksum = uint64(^byte(hexChecksum&0x0F) + 1)
-	binary.LittleEndian.PutUint64(checksumCont, checksum)
-	hex.Encode(hexStringChecksum[:], checksumCont[0:1])
-
-	return hexStringChecksum, nil
 }
